#80 - User 32-bit integers for our herd data IDs

+++ Herd.Business/Models/Commands/GetOAuthURLCommand.cs
@@ -2,7 +2,7 @@
 {
     public class GetOAuthURLCommand : Command
     {
-        public long AppRegistrationID { get; set; }
+        public int AppRegistrationID { get; set; }
         public string ReturnURL { get; set; }
     }
 }
